requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
requests>=2.27.1
pyperclip>=1.8.2